Added empty API_KEY check

+++ tpe.sh
@@ -9,13 +9,19 @@ if [ $# -ne 2 ]; then
   echo "Invalid argument amount. Arguments must be Prefix - Year"
 fi
 
+# Validate API_KEY has been defined
+if [ -z $API_KEY ]; then
+  echo "API_KEY is not defined. Read README.md file before running script."
+  exit 1
+fi
+
 rm *.xml 2> /dev/null || true
 
 echo "Processing league: $1 $2"
 
 
 # Download seasons.xml from API
-curl -s http://api.sportradar.us/rugby-league/trial/v3/en/seasons.xml?api_key=${API_KEY} -o seasons.xml 
+curl -s http://api.sportradar.us/rugby-league/trial/v3/en/seasons.xml?api_key=${API_KEY} -o seasons.xml
 sleep $delay
 # Delete namespace from seasons.xml
 sed -i'' -e 's@xmlns="http://schemas.sportradar.com/sportsapi/rugby-league/v3"@@g' seasons.xml
